backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch==2.4.1
torchvision==0.19.1
timm==1.0.9
pillow==10.4.0
numpy==2.1.1
python-multipart==0.0.9
